+++ src/components/ui/card.tsx
@@ -1,6 +1,6 @@
 import * as React from "react";
 
-import { cn } from "@/lib/utils";
+import { cn } from "../../utils/utils";
 
 function Card({ className, ...props }: React.ComponentProps<"div">) {
   return (
