+++ src/context/store/carouselStore.ts
