+++ src/context/store/sliderStore.ts
