+++ src/pages/Daily/DailyView.tsx
