+++ src/pages/Home/mainDetailPage/DetailContent.tsx
