+++ src/pages/Home/mainDetailPage/DetailContentHeader.tsx
