+++ src/pages/Setting/NameChangePage.tsx
@@ -1,10 +1,10 @@
 import { useNavigate } from "react-router-dom";
 import { ArrowLeft } from "lucide-react";
-import { useState, useEffect } from "react";
-import { BasicInputBox } from "../components/ui/section1/BasicInputBox";
-import BasicButton from "../components/BasicButton";
-import { BasicAlert } from "../components/ui/section1/BasicAlert";
-import BasicPopup from "../components/BasicPopup";
+import { useState } from "react";
+import { BasicInputBox } from "@/components/ui/section1/BasicInputBox";
+import BasicButton from "@/components/BasicButton";
+import { BasicAlert } from "@/components/ui/section1/BasicAlert";
+import BasicPopup from "@/components/BasicPopup";
 
 export function NameChangePage() {
   const navigate = useNavigate();
