--- src/views/index.ts
@@ -1,6 +1,0 @@
-export * from "./main/event_slide_container";
-export * from "./DailyView";
-export * from "./BookshelfView";
-export * from "./WriteView";
-export * from "./GoodsView";
-export * from "./FamilyView";
