+++ tsconfig.tsbuildinfo
@@ -1,0 +1,1 @@
+{"root":["./vite.config.ts","./src/app.tsx","./src/main.tsx","./src/vite-env.d.ts","./src/components/basicbutton.tsx","./src/components/basicpopup.tsx","./src/components/basictoast.tsx","./src/components/bottomnavbar.tsx","./src/components/headerbar.tsx","./src/components/layout.tsx","./src/components/pagecontrol.tsx","./src/components/section2/commentcontent.tsx","./src/components/section2/eventcarousel.tsx","./src/components/section2/imagewithprofiles.tsx","./src/components/section2/inputcontainer.tsx","./src/components/ui/button.tsx","./src/components/ui/card.tsx","./src/components/ui/carouseldetail.tsx","./src/components/ui/carouselmain.tsx","./src/components/ui/dropdown-menu.tsx","./src/components/ui/input.tsx","./src/components/ui/slider.tsx","./src/components/ui/switch.tsx","./src/components/ui/section1/basicalert.tsx","./src/components/ui/section1/basicdropdown.tsx","./src/components/ui/section1/basicinputbox.tsx","./src/lib/utils.ts","./src/lib/api/todoapi.ts","./src/lib/store/carouselstore.ts","./src/lib/store/headerstore.ts","./src/lib/store/sliderstore.ts","./src/lib/store/usecounterstore.ts","./src/pages/index.ts","./src/pages/bookshelf/bookshelfview.tsx","./src/pages/bookshelf/bookshelfviewmodel.tsx","./src/pages/bookshelf/index.tsx","./src/pages/daily/dailyview.tsx","./src/pages/daily/index.tsx","./src/pages/family/familyview.tsx","./src/pages/family/index.tsx","./src/pages/goods/goodsview.tsx","./src/pages/goods/index.tsx","./src/pages/home/memorydetailpage.tsx","./src/pages/home/eventgallery.tsx","./src/pages/home/event_slide_container.tsx","./src/pages/home/homeslider.tsx","./src/pages/home/index.tsx","./src/pages/home/maindetailpage/commentcontainer.tsx","./src/pages/home/maindetailpage/detailcontent.tsx","./src/pages/home/maindetailpage/detailcontentheader.tsx","./src/pages/home/maindetailpage/eventheader.tsx","./src/pages/home/maindetailpage/headernav.tsx","./src/pages/home/maindetailpage/imageslider.tsx","./src/pages/setting/namechangepage.tsx","./src/pages/setting/index.tsx","./src/pages/start/firstmakepage.tsx","./src/pages/start/kakaoinpage.tsx","./src/pages/start/linkinpage.tsx","./src/pages/start/makeconfirmpage.tsx","./src/pages/start/index.tsx","./src/pages/write/writeview.tsx","./src/pages/write/index.tsx","./src/routes/index.tsx","./src/views/index.ts"],"errors":true,"version":"5.7.3"}
